\docType{package}
\name{CausalPairs}
\alias{CausalPairs}
\alias{CausalPairs-package}
\title{Helper functions to load data for the causal pairs effect competition in Kaggle (2013-03)}
\description{
  Helper functions to load data for the causal pairs effect
  competition in Kaggle (2013-03)
}
\author{
  Ahmadou Dicko \email{dicko.ahmadou.gmail}
}
\keyword{package}

